requirements.txt
fastapi
jinja2==3.1.4
uvicorn
python-multipart
python-dateutil
loguru